Print
added a new class of object "Integral" and made Trapezoid and Simpson's inherit its inputs. Modified Integral in order to have an input for rule and created validity tests forcing objects of class trapezoid to have "trapezoid" as their rule input and objects of class simpson to have "simpson" as their rule input. Created methods for integrateIt for both trapezoid and simpson which return the object and the solution to the appropriate integration method. Created a new method for the print function for objects of class integral and within the function defined them as either trapezoids or simpsons depending on the input for "rule," from there applied the appropriate integration function and returned only the result.