Complete
renamed construction packs for Integral, Trapezoid, and Simpson for clarity. Got rid of integration methods because were unneccessary in addition to construction functions I created. Got rid of new generic for print because was extraneous and defined new method based on existing function. Added a workaround to "Simpson" function to fix a bug in which the formula would produce an error if only 3 x values were input. Cleaned up package code and ran checks on all functions.